Most recent observations taken for duplicate demographic data
I found the error for the duplicate in Table 1. One of the people has different ages for what I'm guessing is their Inter-Visit appointments. So they had a birthday between visits.

I will just remove the later one, and use the first date as their age for the Table 1 summary stats. This was done programatically for all data entries